Cleanup: make QVector->float narrowing explicit in AUDAdjustSound
scalepos*pos.i is double (QVector pos); cast to float for the OpenAL position
array. Silences -Wnarrowing.

+++ src/aldrv/al_sound.cpp
@@ -696,7 +696,7 @@ void AUDAdjustSound( const int sound, const QVector &pos, const Vector &vel )
 #ifdef HAVE_AL
     if ( sound >= 0 && sound < (int) sounds.size() ) {
         float p[] = {
-            scalepos *pos.i, scalepos*pos.j, scalepos*pos.k
+            (float)(scalepos * pos.i), (float)(scalepos * pos.j), (float)(scalepos * pos.k)
         }
         ;
         float v[] = {scalevel *vel.i, scalevel*vel.j, scalevel*vel.k};
